handlers(build): handle sparse files in tar
Support for sparse files does not exist in the implementation of
golang tar. This is sad and tracked in [0].

We need this so we should look at supporting it again. This commit
implements a naive version of hole skipping by just seeking over
the "\0" parts of the file if the type of the file is
`tar.TypeGNUSparse`. This copies parts of the io.go::copyBuffer()
code (unfortunately).

[0] https://github.com/golang/go/issues/22735