build: add test for failing osbuild and rework streaming
When adding a test that the error handling is working correctly
the test would fails every 1/10 time with an unexpected EOF error.
Looking deeper it seems the followLineOutput code was racy and
also unnecessary. By just using a wrapped writer and passing
that to `exec.Command.{Stdout,Stderr}` all the heavy lifting
about output handling is done by the exec package. To ensure
that with that we still get streamed output a test is added
that checks that streaming works.

The test also shows that some tweaks are needed inside the handler
(i.e. we cannot use http.Error(), see code comment).