test: simplify/improve TestBuildStreamsOutput
Generate the timestamps of the output from inside the fake osbuild
binary. This way the test that checks if the output is really
streamed can compare the time that the output was generated with
the time the output was recieved. This makes the test a bit
easier to read and understand (I hope, it's still kinda complicated).